Only preserve policy dismissals if targeting default
We show preserved policy dismissals in the vulnerability
report. Vulnerablities are only created for findings
on the default branch. If we preserve dismissals for
MRs that didn't merge into default, we risk showing them
on the vulnerability report even though the dismissal could
not be responsible for introducing the vulnerability to
the default branch.