Remove automatic dictionary generation from `db:migrate`
Generating dictionary files automatically is disruptive, because if the
local database schema has drifted from the branch then this could result
in unstaged changes that need to be restored.

We have validation of the database schema when migrations run, which
means that someone adding a new table with no dictionary file will have
the migrations fail when run locally or in the pipeline. This is
sufficient to catch missing dictionary files and so the automatic
running of dictionary generation should be removed to avoid disrupting
the wider developer community.