Fix email format validator
The validator assumed that both DOM elements, with `.validation-hint`
and `.validation-warning` classes, are used.

While this is true for
`app/views/devise/registrations/_signup_box_form.html.haml`, it is not
for
`ee/app/views/devise/registrations/_signup_box_form_lightweight.html.haml`.

If one or the other class is not present the toggling of `hide` is
ignored and it does not raise any errors.