Drop milestone step from stable branch template
On backports, setting a milestone is done for information purposes, it
is not a required step to proceed with the backport. Lately, assigning
previous milestones has been cumbersome, as the milestone information is
no longer available. This has created friction when backporting bug
fixes into stable branches.

This commit aims to remove the friction by removing the milestone step
from the stable branch template.