Cache results of HasUserType#redacted_name in RequestStore
In some benchmarks for the `/api/v4/projects/:id/jobs` endpoint, we saw
that 18% of the time was spent in `User#redacted_name`. This is because
for every user that is connected to a job, we need to make a number of
queries to check whether the viewing user has permissions to see the
user's name. This had to be done because project bots user the name to
describe the token name, which might leak some useful metadata.

To optimize this a bit, cache the results of the mapping from the
viewing user -> user in RequestStore so we can avoid duplicate work.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/578726

Changelog: performance