Drop cached_route_lookups feature flag
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/57027 was merged
in 2021. While the feature flag has been enabled in staging for a while,
it never was enabled in production.

Given that it never landed in production for this long let's drop it.

We're seeing that this feature flag check is showing up in profiles of
API endpoints that frequently need to generate full paths for
namespaces: https://gitlab.com/gitlab-org/gitlab/-/issues/578726.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/506032

Changelog: changed