Add docs for smartcard hardening recommendations
Currently, there is a risk that users can submit requests to the Rails
monolith with valid certificates of target users, causing the normal
certificate verification by NGINX to be skipped. While we can't remedy
this in the monolith itself, we can recommend users implement security
controls to prevent this attack vector.

Changelog: security