Only use traversal_ids when covered by index
For sub-groups the namespace_traversal_ids perform as good as the old
query. But for root namespaces we need to cover the sorting by an index
to work. So for now we need to conditionally apply the
`namespace_traversal_ids` optimization for root namespaces, when they
are sorted by created or updated_at, for which we have an index.