Update the CommitStatusesFinder to accept ref_type
This allows us to specify which type of refs we're passing in an filter
the pipelines by their respective types. This allows us to be more
accurate when identifying the latest pipeline for each ref.

We also switch to the new latest_pipeline_per_ref method and pass
ref_sha_pairs in. This scope will return the latest pipeline for a sha
within a ref which is required when a sha is within multiple refs i.e.
commit1 could be in master, develop, and tag1.