Update merge request widget to stop polling
Updates the merge request widget to stop the polling requests. Most of,
if not all of, the state changes that happen in the widget are
controlled through the data received from the Apollo subscription so we
are currently just hitting the widget endpoints because of legacy
reasons.

This change however is being done behind a disabled feature flag due to
the impact this change could have if something is missed.