Return correct message when pushing to archived group's project
Previously, when attempting to push to a project whose parent group
was archived, users received a generic "You are not allowed to push
code to this project" message. This was misleading as the actual
reason was the archived status of the parent group.

This commit updates the archive check to use
`self_or_ancestors_archived?` instead of just checking the project's
archived status. Now users will see the more accurate message "You
can't push code to an archived project" regardless of whether the
project itself or its parent group is archived.