Fix rendering old invalid labels records
We found that some labels records have both group_id
and project_id. WE have a constraint to prevent this
and code to fix them when updated. But, some
markdown_cache code was not running callbacks,
so updates were failing on invalid records that needed
to refresh the markdown cache

Changelog: fixed