Make math tests unflaky by cancelling hovers
I've been unable to reproduce the actual cause of the hover, but by
inducing the same hover artificially, I could reproduce the original
test flakes.  This method in turn successfully causes the hover to be
dismissed, and should be stable.

If the tests continue to flake in spite of this, we need to look
at the screenshots to see if the same hover is still appearing.
If so, it suggests *we're* causing it, at some point after the
`create_and_visit_issue_with_description` call.