docs: Move "Redact text" docs to repository_size.md
Currently there were various places that mention the removal of data.
And they were pointing in all directions.

Consolidate docs and put everything together in repository_size.md.
Update all the links to those docs, and keep the headline at the
original location so it can link to the new one.