Update ghost user calls to per-organization model
As part of the preparation for Cells, we must update all internal bot
users to be per-organization rather than cluster-wide. This MR updates
all calls to get the `ghost` user to specify which organization to get
the ghost user for.

Also refactors specs to ensure that all references to
`Users::Internal.ghost` without an organization are removed.

Changelog: changed