Fix database load balancer for newly created OauthAccessToken
On GitLab.com we have an architecture where there is a primary database
and multiple read-only replicas. The replicas have some small amount of
replication lag. We can send read-only queries to the replicas to reduce
traffic to the primary and we do this whenever it is safe.

Today our database load balancing logic relies on some kind of
"sticking" object. Usually the way this works is that when a user
performs a write to the database we "stick" that user to the primary.
This generally works for users that are authorized with cookies because
subsequent requests include the user_id and we can look up exactly what
their "sticking" state is.

But it does not work for newly generated OAuth tokens. As discovered in
https://gitlab.com/gitlab-org/gitlab/-/issues/579054#note_2857344904 we
have an issue where we don't actually know which user the OAuth token
belongs to until we load the OAuth token from the database. But if we
load the token from a replica then it might not exist.

Our load balancer does already have to deal with similar edge cases
elsewhere as in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/136422 . The way
we do this is by introducing a new sticking object. In this case we take
a secure non-reversible hash of the OAuth token and store that as the
sticking key in Redis. Then in future when we get a request with an
OAuth token we can look up in Redis to see if it has a sticking session.
If so we use the primary (or a sufficiently caught up replica for this
specific sticking LSN position).