Update access token expiration webhooks ui
The UI for access token expiration webhooks is inconsitent
between projects and groups, and should be aligned. We also
don't need links to docs that simply explain how to get back
to the settings page you are already at. Finally, the project
setting seems to be blocked from rendering due to leftover
feature flag code, so this removes that code.

Changelog: fixed
EE: true