Raise errors in fixed items model methods
Raise errors if the passed attribute doesn't
exist for the model class like
Active Records handles it.

This affects the methods .where and .find_by

Also raise an error when using find and no item
was found with the given id.

This is done to reach parity with how Active Records
handles these situations.