Allow specifying ref_type while resolving pipelines
Ci::Pipeline objects have a `tag` attribute which specifies if the
pipeline was triggered for a tag ref.

This change adds a new argument to the ResolvesPipelines module which
allows users to pass a ref type enum value of either heads, tags, or
omit it.

The pipelines finder allows us to filter by a scope. These scopes
include both branches and tags. The resolver will convert `ref_type:
'heads'` to `scope: 'branches'`, and `ref_type: 'tags'` to `scope:
'tags'`.

This allows us to filter the pipelines by the appropriate type.