Retain format when redirecting moved projects
While accessing the API we check if the project has been moved and
redirect to the same path with the project's old full path exchanged
with the project's ID.

Previously this logic would not retain the format segment of the URL
i.e. https://gitlab.com/api/v4/project/moved/repositories/archive.zip
would redirect to
https://gitlab.com/api/v4/project/moved/repositories/archive omitting
the `.zip` extension.

This change fixes the logic to ensure we include the format when
forwarding.

Fixes https://gitlab.com/gitlab-org/gitlab/-/issues/578087+