Fetch correct pipeline when viewing commits in detacted state
If you are viewing the commits and have a branch or tag selected in the
ref selector we can correctly fetch the pipeline using the ref.name.
However, you can also view commits when you've checked out a commit i.e.
instead of viewing commits/master you're viewing commits/SHA and viewing
all of the commits in the history of SHA.

When we are in this detatched state we need to remove the ref and rely
on the SHA instead. We don't know which branch or tag we're on so we
just find the most recent pipeline within any ref and show that one.

Unfortunately we don't have ref_type available in the commits page so
here we are checking if the ref looks like a SHA. This does have a
caveat that any branches that look like SHAs will show the latest
pipelines for the commit, not the commit within the current branch.

This trade-off is much better than the current situation. It's highly
unlikely for users to be naming branches with SHA values. It is much
more likely for users to view the commits page while viewing a commit.