Fix NoMethodError when current_user is nil in MR subscription
Contributes to https://gitlab.com/gitlab-org/gitlab/-/work_items/579309

**Problem**

When `current_user` is nil in the
`Subscriptions::User::MergeRequestUpdated` GraphQL subscription, the
authorized? method was attempting to call .id on nil, causing a
NoMethodError.

**Solution**

Use the safe navigation operator (&.) to handle the case
when current_user is nil, allowing the authorization check to properly
fail and unsubscribe the user instead of raising an exception.

Changelog: fixed