Clarify similar changeset-methods in pipeline model
Recently a new `#changed_paths` method was introduced that serves a
similar purpose as the existing `#modified_paths` method, but behaves
differently in a meaningful way that is not obvious.

This adds more explicit comments highlighting the differences to avoid
the risk of confusion.

See https://gitlab.com/gitlab-org/gitlab/-/issues/578368