Adapt Puma config to support Puma 7
The Puma configuration template needs to be adjusted to work with Puma 7:

1. The callback hooks have been renamed in https://github.com/puma/puma/pull/3438:
   - `on_worker_boot` -> `before_worker_boot`
   - `on_worker_shutdown` -> `before_worker_shutdown`
   - `on_restart` -> `before_restart`

This commit adds a conditional block based on the Puma version to use
the new hooks.

2. With https://github.com/puma/puma/pull/3297, Puma 7 raises an
exception if Puma's options are accessed before they are fully
loaded. Currently we pass these options during loading of the Puma
config file, which then triggers this exception. These options are used
to determine whether Puma is in clustered mode by checking the number of
workers.

This commit uses the `set_puma_worker_count` method added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/204104.

Related changes:
- Omnibus: https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/8715
- CNG: https://gitlab.com/gitlab-org/build/CNG/-/merge_requests/2643