Include support for when grape transforms headers
Currently we're looking at headers after they are parsed by rails or grape,
which leads to slight inconsistencies when trying to access them. This
change aims to read the values from request.env (which is a rack request env)
which allows us to read the header values in a consistent manner.