Update handling of isFetching as loading state
Set isFetching to be the opposite of the loading state using the
suggestionsLoading handler.

This will allow the showNoMatchesText to properly appear when the
loading process has finished.
It will also be used to show the loading wheel at the time of the
creation of the list of suggestions. This will ensure that tokens
without initialSuggestions avoid showing a brief flash of empty before
the loading icon appears.