Use blob presenter method instead of redefining
The blob type in graphql has a field raw_text_blob which specifies it
uses the method `text_only_data` which doesn't exist. While trying to
figure out how this works I realised that graphql was using the
`raw_text_blob` method defined in the same class. I also noticed that
this method is redefining what is already defined in the blob presenter.

This change removes the raw_text_blob method and uses the correct
presenter method for returning the raw plain data.