Allow instance admins to hard-delete AI items
Provides a `forceHardDelete` argument for AI Catalog item mutations that
can only be used by an instance admin. It will always hard-delete the
item, and never soft-delete. This is a measure of security in the case
where something must be deleted and any active uses of the item must
also be deleted.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/575127