Create the `new_ui_enabled` column
This creates the `new_ui_enabled` column in the `user_preferences`
table. This will replace the `project_studio_enabled` column which
wasn't flexible enough for us to roll the new UI out to .com users as it
wasn't nullable.

In a follow-up change, we'll update the product logic to leverage this
new column to determine whether or not the new UI should be presented to
the user.