Ignore exclude_projects when using traversal_ids
The exclude_projects param was an optimization when querying epic work
items. However, this optimization is not necessary and counterproductive
when we use traversal_ids to query the work item hierarchy.

We therefore ignore the parameter when the feature flag is enabled so we
can rollout the feature flag for the current epic lists.