Render correct pipeline statuses for branches
When viewing the branches page for a project, the pipeline status icon
for each branch points to the latest pipeline for a ref matching the
branch's name, however, if the project also includes a tag with the same
name and a pipeline ran after the one for the branch the incorrect
pipeline status is displayed.

Now we filter by pipelines for branches so the status will always be the
correct one.

Changelog: fixed