Allow re-adding to merge train when pipeline fails
Previously, when a merge train pipeline failed in a project with
"Pipelines must succeed" enabled, one had to manually run another
pipeline before being able to retry the merge. This requirement was not
obvious. Nor was it useful, given that the worst that can happen is that
the pipeline fails again (causing the train to restart).

This change is behind the feature flag `allow_merge_train_retry_merge`,
disabled by default.

See https://gitlab.com/gitlab-org/gitlab/-/issues/451302

EE: true