Ensure Duo Developer service account membership is synced immediately
As discovered in
https://gitlab.com/gitlab-org/gitlab/-/issues/578824#note_2870763994 we
have an issue when we use Duo Agent Platform for the first time within a
project. In `::Ai::ServiceAccountMemberAddService#execute` we are
giving this service account Developer role in the project the first time
it's used in a project.

This causes predictable issues with permission checks because the
immediate next step for remote flows is to create a branch. When we
create a branch we are checking that the user has permission to the
project.

We are checking it in 2 ways:

1. Explicitly in `WorkloadBranchService`
2. Implicitly after using Gitaly to trigger the branch creation it then
   calls back to GitLab Rails to confirm the user has authorization to
   create this branch

A few days ago we were failing at (2) but since
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/209783 was merged
we actually fail earlier now at (1).

The cause of this problem is that permission creation has always been
asynchronous in GitLab because our permissions are always synced over to
a separate `ProjectAuthorization` model. This `ProjectAuthorization` is
what we use for checking access and not the `Member` model that we just
created. The asynchronous update is triggered in ActiveRecord callbacks
at
https://gitlab.com/gitlab-org/gitlab/-/blob/08904ec881f57f992056aa391d199895e6968b18/app/models/member.rb#L376

This MR fixes the problem by immediately triggering the creation of
`ProjectAuthorization` for this user and project combination as soon as
we create the membership.