Fix ResumeWorker not scheduling to the correct queue
`ResumeWorker` execution with an worker name argument will try to
reschedule another `ResumeWorker` job with 1s delay.

However, using `.perform_in` would route the job to `ResumeWorker`'s
queue which is the default/catchall queue. Instead, we want to run the
`ResumeWorker` in the same queue that the resumed worker is running.

The fix is by using the Sidekiq API `Sidekiq::Client.enqueue_to_in`.