Fix root_path check & project parsing in SM audits
Prior to this change, we had implemented the logic
to check if the auth token file path within the Rails root
directory based on local GDK logic. It is different for
production containers including CNG and Omnibus. This was
causing error when validating the openbao Rails
Auth token file path is validating. Also, the structure of JSON audit
log payload was changed a bit after we added the namespacing in Openbao

This change fixes the issue by updating the logic
to match correctly the Rails app root path in
production environments. Also, we updated the JSON parsing logic to get
the project details based on the changes introduced due to namespacing

[Discussion]
Thread
(https://gitlab.com/gitlab-org/charts/gitlab/-/merge_requests/4549#note_2869604889)

Changelog: changed
EE: true