Force work items actions to be right-aligned
The work items actions dropdown was always meant to be right-aligned,
but was using a deperacted prop which ultimately had no effect on the
dropdown's positioning.

This updates the prop accrodingly.

This should also fix rare edge-cases where the dropdown could be
left-aligned in cases where there is no room for it to render, causing
the content to be truncated.

Changelog: fixed