Disable Workhorse Rails skip interval in readiness check by default
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/208379, we
defaulted the value of `rails_skip_interval` 20 seconds to signal that
Workhorse did not need to issue a `/-/readiness` if it proxied a
successful request to the Rails backend in the last 20 seconds. The idea
is that this would make the readiness checks less prone to flapping, and
Workhorse would not need to send a `/-/readiness` probe to Puma as
frequently.

However, in practice, we found on a heavily loaded Puma server, the
`/-/readiness` probe served as a load shedding mechanism. Without this,
it's possible for request queue durations to increase to the point where
the liveness checks start to fail. This causes unintended 502 errors
because a pod failing a liveness check will be terminated immediately.

Disable the `rails_skip_interval` setting by default for now since we
want to be able to enable the Workhorse readiness checks in production,
and this default matches the current readiness probe behavior.

Relates to
https://gitlab.com/gitlab-com/gl-infra/production-engineering/-/issues/27899

Changelog: changed