Stop unescaping HTML in BaseLabel#title=, #description=, Timebox#title=
This affects anything including BaseLabel or Timebox, namely Label,
AntiAbuseReports::Label, Milestone, and Iteration.

## Why?

Calling `CGI.unescapeHTML` immediately after `Sanitize.clean` is an
anti-pattern. Here's an example of `Sanitize.clean` removing something
nasty:

```irb
[83] pry(main)> Sanitize.clean("hello <script>alert('xss')</script>")
=> "hello "
```

What happens if the user enters entities instead?

```irb
[84] pry(main)> Sanitize.clean("hello &lt;script&gt;alert('xss')&lt;/script&gt;")
=> "hello &lt;script&gt;alert('xss')&lt;/script&gt;"
```

Entities are safe, so this is all OK!  But what if we then `CGI.unscapeHTML` it?

```irb
[85] pry(main)> CGI.unescapeHTML(Sanitize.clean("hello &lt;script&gt;alert('xss')&lt;/script&gt;"))
=> "hello <script>alert('xss')</script>"
```

Eep!  This is **the current state of the application**; we are
essentially _not_ sanitising the output, because we're allowing the
user to enter anything in entities and then unescaping it *after*
sanitisation.

Here's an example demonstrating this, but you can easily test it
yourself by going to create a new label through the web interface.  SQL
logs output by the console have been omitted for clarity.

```irb
[96] pry(main)> l = Label.first
=> #<ProjectLabel id:80 gitlab-org/gitlab-test~80>

[99] pry(main)> l.title = "hello &lt;script&gt;alert('xss')&lt;/script&gt;"
=> "hello &lt;script&gt;alert('xss')&lt;/script&gt;"

[100] pry(main)> l.save!
=> true

[101] pry(main)> l.reload
=> #<ProjectLabel id:80 gitlab-org/gitlab-test~80>

[102] pry(main)> l.title
=> "hello <script>alert('xss')</script>"
```

The good news is, sanitisation is not the answer here!  The user should
be able to enter any text as a label or milestone title, and that
text should be shown exactly as entered --- no sanitisation needed.
We must treat it as text, not HTML.  Thankfully, we should be doing
this already for the most part, seeing as (as mentioned above) we are
already allowing users to enter non-sanitised content here via entity
unescaping.