Render missing HEAD message for tags which do not reference a commit
A tag can reference a tree or blob object. In this case the UI shows the
default branch's HEAD commit because our `commit` method uses root_ref
in place of a nil ref argument. This is a deliberate choice and works
for most situations but for tags it causes this bug.

Now when a tag does not include a target commit we display a message
that we cannot find the HEAD of the tag.

Changelog: fixed