Render correct pipeline statuses for tags
When viewing the tags page for a project, the pipeline status icon
for each tag points to the latest pipeline for a ref matching the
tag's name, however, if the project also includes a branch with the same
name and a pipeline ran after the one for the tag the incorrect
pipeline status is displayed.

Now we filter by pipelines for tags so the status will always be the
correct one.

Changelog: fixed