Return error if failing to create WorkflowsWorkload
This is a minor refactor that shouldn't have any impact on behaviour
today. This code path which was previously ignoring the return value of
`#create` is slightly risky in Rails because it means errors might just
be hidden.

It's not possible for this to error today given how the code is written
but having this error handling in place makes it a bit more robust to
changes in future. It also makes it easier for developers reading this
code to see that this class does correctly handle all errors.