Implement to_param on Integrations::Prometheus
The removal of the code in this class caused an issue when viewing the
integration list for a project which previously had the prometheus
integration enabled.

Project#find_integration does the following:

```ruby
def find_integration(integrations, name)
  integrations.find { _1.to_param == name }
end
````

Calling `to_param` on the `Integrations::Prometheus` record caused the
`NotImplementedError` on the base concern to trigger.

This MR re-implements the Prometheus methods which "must" exist in
order to avoid the not-implemented method being called.

Changelog: fixed