Delete orphaned clusters and add sharding key not null constraint
Cluster records created prior to https://gitlab.com/gitlab-org/gitlab/-/merge_requests/198239
were not removed when their parent project/group was deleted as
there was no foreign key relationship between the two tables.

This merge request uses a post deployment migration to delete all
records that were left in this orphaned state before the fix was
implemented. We need to remove these records so that we can then
add a NOT NULL constraint on project_id/group_id/organization_id,
which is required for these to be used as the sharding key(s).

Changelog: other