Doc(components): add access_token to curl call
The job example that calls the GitLab API to check
whether the job a component adds was properly created
makes the assumption that the comoonent project is public.
Thus, the job will fail in all other cases with a 404
error. Adding the job token as query parameter ensures that
the API can be reach.