Use adapter for EpicIssues::DestroyService
This adds a new adapter for the EpicIssues::DestroyService to always use
the WorkItem service to destroy EpicIssues.

We can also remove the code that deletes the EpicIssue record, since
this gets now handled via the foreign key constraint.

No change in the API since the data was bi-directionally synced already.