Fix Gitaly N+1 in /api/v4/project/:id/jobs (behind FF)
There's a gitaly N+1 in the jobs API. The `commit` field is retrieved
via `Commit.lazy`, but only at serialization time, which forces the
eager evaluation of each individual commit. We fix this by evaluating
the `commit` method of each build in the paginated result set prior to
serialization.

Given the criticality of the endpoint, this fix is behind the feature
flag `fix_jobs_api_gitaly_n_plus_1` (disabled by default), to derisk the
rollout.