Duo Agent Platform only use instance or top level group runners
More motivation can be found in
https://gitlab.com/gitlab-org/gitlab/-/issues/578791 .

This MR changes things so that Duo jobs are only able to run on instance
wide or top level group runners.

It does this in 2 parts:

1. It adds the `gitlab--duo` tag to all Duo jobs. This means that only
   runners tagged with `gitlab--duo` will run these jobs. We will update
   our instance wide runners to have these tags. And customers can add
   this tag to their top-level group runners. This is behind a feature
   flag `duo_agent_platform_ci_job_tags`.
2. It actively fails any job that is assigned to a runner that isn't an
   instance or top level group runner. This means that if someone adds a
   `gitlab--duo` tag to a project runner (for example) we
   will still not run the job on that runer. An error will be displayed
   telling them to remove the tag. This is behind another feature flag
   `duo_runner_restrictions`.

The runner blocking logic behind a feature flag and the intention is to only
enable this on GitLab.com as it isn't that useful for Self-Managed.

On GitLab.com we already have the `gitlab--duo` tag applied to our small
amd64 runners.

Changelog: changed