Allow labels table to be marked as sharded
Issue to validate constraint: https://gitlab.com/gitlab-org/gitlab/-/issues/558353

Following https://gitlab.com/gitlab-org/gitlab/-/merge_requests/202429.

This change allows for the regex looking for a NOT VALID constraint defined to be be
matched and to mark the table as sharded now that we also verify the column 'convalidated'.