Fix MR approver list with composite identity filtering
Approval list include anyone who has commented on the MR, and uses
ability checks to decide if they are eligible approvers for the change.
But ability checks perform composite identity filtering that can
transform the user to `nil`.

We may need to re-think that approach (and soon!), as it is bound to
impact all sorts of backend logic not expecting to ever have an already
loaded user disappear (especially Sidekiq jobs). But to solve the
immediate problem, we make the affected ability check robust to `nil`
users.

See https://gitlab.com/gitlab-org/gitlab/-/issues/580396

Changelog: fixed