Fix export_as_csv_spec.rb QA test
The legacy issues list no longer renders in GitLab - the work
items list renders instead. export_as_csv_spec.rb never fully
execute because `if has_element?('issues-list-more-actions-dropdown')`
which looks for an element in the legacy issues list is never found.
We are updating it here so it executes against the work items list
instead.