Add FF to eagerly resume jobs
`ResumeWorker` by default will resume a batch of jobs at a time. This
adds an ops FF `concurrency_limit_eager_resume_processing` that tries to resume
as many batches of jobs as possible in 5 minutes.

However this comes with a risk - there is an edge case when
sidekiq queue is massively backlogged,
the number of concurrent jobs of a worker could far exceed
the set concurrency limit. When this happens, it could put a lot of
pressure to the database.

The FF `concurrency_limit_eager_resume_processing` is meant to be
enabled for
GitLab.com as the performance of resuming jobs is important especially
for .com.

The FF is disabled by default, so self-managed and Dedicated instances
will revert to the previous behaviour of resuming a batch of jobs at a
time, to prevent the impact their database.

Changelog: other