Add convenience script to prune local branches
As we regularly squash commits when merging merge requests it's not
straightforward to figure out which branches have been merged. Normally
it would be as simple as using `git branch -d` and allow it to figure
out which commits exist in the base branch before deleting the branches.

This script uses the glab cli tool to check if there is a merged MR with
the branch as the source branch. If it finds a merged MR then we know
we can force delete this branch.