Fix incorrect enrolment of SSO users in Email-based OTP
We have a feature to enrol new users in Email-based OTP when they sign
up, controlled by a Feature Flag.

Previously, any new user would be enrolled if the FF was enabled. This
is incorrect, as Email-based OTP only applies to users who sign in to
GitLab with a password. When users solely use their own identity
provider, such as via SSO, SAML, LDAP, etc, then they are out of scope.

This fix checks whether "password_automatically_set" is true. It is
true when the user has no ability to log into GitLab with a password,
and in such cases the user should NOT be enroled in Email-based OTP.

Fixes https://gitlab.com/gitlab-org/gitlab/-/issues/580743