Temporarily reduce authTokens validation requirements
We're updating the authentication process for the observability app, and
an interim change is to reduce the validation requirements for the
authTokens prop so that when we update the service and it returns
less properties, we don't break the app.

This will be reverted in a follow-up PR after we update the service to
return the required properties.