Add feature flag check to vulnerability_reads unset triggers
This adds the vulnerability_management.dont_execute_db_trigger feature
flag check to two database triggers that were missing it:

- unset_has_issues_on_vulnerability_reads (triggered on
  vulnerability_issue_links deletes)
- unset_has_merge_request_on_vulnerability_reads (triggered on
  vulnerability_merge_request_links deletes)

These triggers handle the unsetting of has_issues and has_merge_request
flags when the corresponding links are deleted. They need to respect
the same feature flag as the other vulnerability_reads triggers to
ensure consistency when comparing state between projects with and
without the feature flag enabled.

This is a follow-up to
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/212278 which added the
feature flag check to the corresponding "set" triggers.

Changelog: changed