Change the urgency of Ci::CreateDownstreamPipelineWorker
According to the urgency documentation, a job should be marked as
"urgent" (high) only if users perceive delays as bugs. In this case,
downstream pipeline creation happens in the background while the parent
pipeline is running. Users don't immediately notice these delays,
so this doesn't meet the criteria for high urgency.

Changelog: other