Remove dangling scans even if there are no scannable sboms
There is a flaw in the `remove_dangling_dependency_scans` logic where it
is only executed if the pipeline contains at least one SBoM with
dependency_scanning results. This means that if none of the SBoMs can be
used to generate a security report, we end up with scans stuck in the
created state. This change fixes the issue by always removing dangling
scan recoreds.

Changelog: fixed
EE: true