Always show page size selector on vulnerability report
This makes sure that the page size selector is always shown on the
vulnerability report. If there were to be an error, especially with
query complexity, the user can change the page size to a lower amount
so the query is less complex and it may succeed.

Changelog: changed
EE: true