Allow failed merge train merges to be retried
Prior to this change, when "Pipelines must succeed" was enabled, merge
train pipeilne failures required a new merge request pipeline to succeed
before the MR could be re-added to the train. This added unneccesary
friction. For example: The failure could have been inherited from the
target branch, or an MR earlier on the train. The failure could also
have been a flaky test. In any case, re-adding to the train is a
low-risk operation: In the worst case, the pipeline fails again. There
is a cost to that, which is why we have a warning modal but empower the
merge user to make the right call.

This behaviour is behind the feature flag
`allow_merge_train_retry_merge`, as of this commit enabled by default.

- Rollout issue: https://gitlab.com/gitlab-org/gitlab/-/issues/580323
- Feature issue: https://gitlab.com/gitlab-org/gitlab/-/issues/451302

Changelog: changed
EE: true