Set required SlackIntegration relation in specs
This commit is to lay the groundwork for a future MR in which we add
the multi-column not-null constraint to the Slack Integrations sharding
key.

By ensuring adding an organization/group traits to the factory now, the
next MR should be easier.

It also updates the BulkUpdateService to ensure the sharding key is set
when re-creating missing integrations.