Fix /admin/sidekiq not loading CSS assets in Cloud Native GitLab
In GitLab 14.3, we added a workaround in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/70113 to disable
the Sendfile interface for Sidekiq assets. This workaround deleted the
HTTP_X_SENDFILE_TYPE header, which signaled to Rack not to use the
Sendfile interface.

Sidekiq bundles its own CSS and JavaScript assets in the gem, and
registers these via `Rack::Static` By default, `Rack::Sendfile` will use
the Sendfile interface if available to offload the sending to Workhorse
However, Workhorse might not have these assets, as they are bundled with
the installed Ruby gem.

However, with the upgrade to Rack v2.2.20+, HTTP_X_SENDFILE_TYPE is
ignored. Rails will always use the Sendfile interface if the body
responds to `to_path`. In this case, `Rack::Static` uses `Rack::Files`,
which responds to `to_path`.

To work around this issue, we remove that header and read the file
content directly to restore the body that Rack::Sendfile emptied.

Relates to
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/20862

Changelog: fixed