Add warning alert and beta badge to new security dashboard
This adds a warning alert about the no longer detected vulnerabilities
that are include in the vulnerabilities over time chart and a beta badge
to the title.

The no longer detected warning alert is guarded by the same logic as
the information auto resolve alert, albeit with different localstorage
keys.